Improve code consistency
