+++ client/dashboard/app/queries/domains.ts
@@ -1,7 +1,5 @@
-import { mutationOptions, queryOptions } from '@tanstack/react-query';
-import { fetchDomains, fetchDomainSuggestions, updateDNSSEC } from '../../data/domains';
-import { queryClient } from '../query-client';
-import { siteDomainsQuery } from './site-domains';
+import { queryOptions } from '@tanstack/react-query';
+import { fetchDomains, fetchDomainSuggestions } from '../../data/domains';
 import type { DomainSuggestionQuery } from '../../data/types';
 
 export const domainsQuery = () =>
@@ -17,12 +15,4 @@ export const domainSuggestionsQuery = (
 	queryOptions( {
 		queryKey: [ 'domain-suggestions', search, domainSuggestionQuery ],
 		queryFn: () => fetchDomainSuggestions( search, domainSuggestionQuery ),
-	} );
-
-export const updateDNSSECMutation = ( domain: string, siteId: number ) =>
-	mutationOptions( {
-		mutationFn: ( enabled: boolean ) => updateDNSSEC( domain, enabled ),
-		onSuccess: () => {
-			queryClient.invalidateQueries( siteDomainsQuery( siteId ) );
-		},
 	} );
