+++ client/dashboard/data/domains.ts
@@ -21,45 +21,16 @@ export interface Domain {
 		status: string;
 	};
 	expiry: string | false;
-	is_dnssec_enabled: boolean;
-	is_dnssec_supported: boolean;
 	is_eligible_for_inbound_transfer: boolean;
 	is_hundred_year_domain: boolean;
 	is_wpcom_staging_domain: boolean;
+	is_dnssec_enabled: boolean;
+	is_dnssec_supported: boolean;
 	primary_domain: boolean;
 	site_slug: string;
 	type: `${ DomainTypes }`;
 	wpcom_domain: boolean;
 }
-
-export interface DNSSECDSData {
-	rdata: string;
-	key_tag: number;
-	algorithm: number;
-	digest_type: number;
-	digest: string;
-}
-
-export interface DNSSECDNSKey {
-	rdata: string;
-	flags: number;
-	protocol: number;
-	algorithm: number;
-	public_key: string;
-}
-
-export interface DNSSECCryptokey {
-	dnskey: DNSSECDNSKey;
-	ds_data: DNSSECDSData[];
-}
-
-export type DNSSECResponse = {
-	data?: {
-		cryptokeys: DNSSECCryptokey[];
-		dnssec_data_set_at_registry: boolean;
-	} | null;
-	status: string;
-};
 
 export async function fetchDomains(): Promise< Domain[] > {
 	const { domains } = await wpcom.req.get( '/all-domains', {
@@ -94,12 +65,4 @@ export async function fetchDomainSuggestions(
 	);
 
 	return suggestions;
-}
-
-export async function updateDNSSEC( domain: string, enabled: boolean ): Promise< DNSSECResponse > {
-	return wpcom.req.post( {
-		path: `/domains/dnssec/${ domain }`,
-		apiNamespace: 'wpcom/v2',
-		...( ! enabled && { method: 'DELETE' } ),
-	} );
 }
